data-sync/requirements.txt
psycopg2-binary==2.9.7
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.0
sqlalchemy==2.0.21
alembic==1.12.0
pydantic==2.4.2
structlog==23.1.0
tenacity==8.2.3